Give `Var`s an `aval`.
